Switched to MySQL + migrations added + refactoring
