+++ .gitignore
@@ -3,4 +3,3 @@ obj/
 /packages/
 .DS_Store
 .idea
-appsettings.json
