--- UltraMafia.csproj
@@ -1,38 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
-        <LangVersion>8</LangVersion>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <Compile Remove="UltraMafia.DAL\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <EmbeddedResource Remove="UltraMafia.DAL\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Remove="UltraMafia.DAL\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Content Remove="UltraMafia.DAL\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <_ContentIncludedByDefault Remove="UltraMafia.DAL\obj\project.assets.json" />
-      <_ContentIncludedByDefault Remove="UltraMafia.DAL\obj\project.packagespec.json" />
-      <_ContentIncludedByDefault Remove="UltraMafia.DAL\obj\UltraMafia.DAL.csproj.nuget.dgspec.json" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="UltraMafia.DAL\UltraMafia.DAL.csproj" />
-    </ItemGroup>
-    <ItemGroup>
-        <PackageReference Include="Telegram.Bot" Version="15.3.0" />
-    </ItemGroup>
-
-</Project>
