+++ src/UltraMafia.DAL/Model/GameSessionMember.cs
