+++ src/UltraMafia/Frontend/IFrontend.cs
