+++ src/UltraMafia/GameModel/ApproveVoteDescriptor.cs
