+++ src/UltraMafia/GameModel/VoteDescriptor.cs
