+++ src/UltraMafia/GameService.cs
@@ -11,6 +11,7 @@ using UltraMafia.DAL.Enums;
 using UltraMafia.DAL.Model;
 using UltraMafia.Frontend;
 using UltraMafia.GameModel;
+using UltraMafia.Helpers;
 
 namespace UltraMafia
 {
