+++ src/UltraMafia/Helpers/GameSettings.cs
@@ -1,7 +1,8 @@
-namespace UltraMafia
+namespace UltraMafia.Helpers
 {
     public sealed class GameSettings
     {
         public int MinGamerCount { get; set; } = 4;
+        public bool DevelopmentMode { get; set; } = false;
     }
 }
