+++ src/UltraMafia/UltraMafia.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+    <PropertyGroup>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <LangVersion>8</LangVersion>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <None Include="settings.json">
+            <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\UltraMafia.DAL\UltraMafia.DAL.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Telegram.Bot" Version="15.3.0" />
+    </ItemGroup>
+
+</Project>
