+++ src/UltraMafia/settings.json
@@ -1,0 +1,17 @@
+{
+  "Db": {
+    "Host": "<db-host>",
+    "Port": 3306,
+    "User": "<db-user>",
+    "Password": "<db-password>",
+    "DbName": "UltraMafia"
+  },
+  "Frontend": {
+    "Token": "<bot-token>",
+    "BotUserName": "<bot-username>"
+  },
+  "Game": {
+    "MinGamerCount": 4,
+    "DevelopmentMode": false
+  }
+}
